+++ web/modules/custom/qs_badge/src/Service/BadgeManager.php
@@ -4,6 +4,7 @@ namespace Drupal\qs_badge\Service;
 
 use Drupal\Core\Session\AccountProxyInterface;
 use Drupal\Core\Database\Connection;
+use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\taxonomy\TermInterface;
 use Drupal\user\UserInterface;
 use Drupal\Core\Datetime\DrupalDateTime;
@@ -27,12 +28,21 @@ class BadgeManager {
    */
   protected $connection;
 
+
+  /**
+   * The node Storage.
+   *
+   * @var \Drupal\node\NodeStorageInterface
+   */
+  protected $nodeStorage;
+
   /**
    * Class constructor.
    */
-  public function __construct(AccountProxyInterface $currentUser, Connection $connection) {
+  public function __construct(AccountProxyInterface $currentUser, Connection $connection, EntityTypeManagerInterface $entity_type_manager) {
     $this->currentUser = $currentUser;
     $this->connection = $connection;
+    $this->nodeStorage = $entity_type_manager->getStorage('node');
   }
 
   /**
@@ -75,6 +85,66 @@ class BadgeManager {
     }
 
     return $events_by_day;
+  }
+
+  /**
+   * Get events for every days between two dates where user has subscriptions.
+   *
+   * @param \Drupal\taxonomy\TermInterface $community
+   *   The community entity.
+   * @param \Drupal\Core\Datetime\DrupalDateTime $date_start
+   *   The start date.
+   * @param \Drupal\Core\Datetime\DrupalDateTime $date_end
+   *   The end date.
+   * @param bool $status
+   *   The required status for the subscriptions.
+   * @param \Drupal\user\UserInterface $account
+   *   The user entity.
+   *
+   * @return array[]
+   *   A collection of dates where it occur an event with subscriptions.
+   *   Otherwise an empty array.
+   */
+  public function getSubscriptionByDates(TermInterface $community, DrupalDateTime $date_start, DrupalDateTime $date_end, $status = TRUE, UserInterface $account = NULL) {
+    $user = $this->currentUser;
+    if (!is_null($account)) {
+      $user = $account;
+    }
+
+    $query = $this->connection->select('node_field_data', 'event');
+    $query->fields('event', ['nid'])
+      ->condition('event.type', 'event')
+      ->condition('event.status', TRUE);
+
+    $query->leftJoin('node__field_activity', 'field_activity', 'field_activity.entity_id = event.nid');
+    $query->leftJoin('node__field_community', 'field_community', 'field_community.entity_id = field_activity.field_activity_target_id');
+    $query->condition('field_community.field_community_target_id', [$community->id()], 'IN');
+
+    $query->leftJoin('node__field_start_at', 'field_start_at', 'field_start_at.entity_id = event.nid');
+    $query->condition('field_start_at.field_start_at_value', [$date_start->format('c'), $date_end->format('c')], 'BETWEEN');
+
+    $query->leftJoin('subscriptions', 'subscriptions', 'subscriptions.entity = event.nid');
+    $query->condition('subscriptions.user', $user->id());
+
+    if ($status) {
+      $query->condition('subscriptions.status', $status);
+    }
+    else {
+      $query->condition('subscriptions.status', NULL, 'is');
+    }
+
+    $query->addExpression("DATE_FORMAT(field_start_at.field_start_at_value, '%Y-%m-%d')", 'formated_day');
+
+    $query->orderBy('formated_day', 'ASC');
+
+    $rows = $query->execute()->fetchAll();
+
+    $subscriptions_events_by_day = [];
+    foreach ($rows as $row) {
+      $subscriptions_events_by_day[$row->formated_day][$row->nid] = $this->nodeStorage->load($row->nid);
+    }
+
+    return $subscriptions_events_by_day;
   }
 
   /**
@@ -278,6 +348,69 @@ class BadgeManager {
     $privileges = [];
     foreach ($rows as $row) {
       $privileges[$row->entity][$row->privilege] = $row->privilege;
+    }
+
+    return $privileges;
+  }
+
+  /**
+   * Get the list privilege for the user for every days between two dates.
+   *
+   * The user's privileges are ordered from lowest to highest.
+   *  - activity_members,
+   *  - activity_maintainers,
+   *  - activity_organizers.
+   * It means you have the highest privilege in the last index of the array.
+   *
+   * @param \Drupal\taxonomy\TermInterface $community
+   *   The community entity.
+   * @param \Drupal\Core\Datetime\DrupalDateTime $date_start
+   *   The start date.
+   * @param \Drupal\Core\Datetime\DrupalDateTime $date_end
+   *   The end date.
+   * @param \Drupal\user\UserInterface $account
+   *   The user entity.
+   *
+   * @return array[]
+   *   A collection of dates where it occur an event with subscriptions.
+   *   Otherwise an empty array.
+   */
+  public function getPrivilegesByDates(TermInterface $community, DrupalDateTime $date_start, DrupalDateTime $date_end, UserInterface $account = NULL) {
+    $user = $this->currentUser;
+    if (!is_null($account)) {
+      $user = $account;
+    }
+
+    $query = $this->connection->select('node_field_data', 'event');
+    $query->fields('event', ['nid'])
+      ->condition('event.type', 'event')
+      ->condition('event.status', TRUE);
+
+    $query->leftJoin('node__field_activity', 'field_activity', 'field_activity.entity_id = event.nid');
+    $query->leftJoin('node__field_community', 'field_community', 'field_community.entity_id = field_activity.field_activity_target_id');
+    $query->condition('field_community.field_community_target_id', [$community->id()], 'IN');
+
+    $query->leftJoin('node__field_start_at', 'field_start_at', 'field_start_at.entity_id = event.nid');
+    $query->condition('field_start_at.field_start_at_value', [$date_start->format('c'), $date_end->format('c')], 'BETWEEN');
+
+    $query->leftJoin('privileges', 'privileges', 'privileges.entity = field_activity.field_activity_target_id');
+    $query->fields('privileges', ['privilege'])
+      ->condition('privileges.user', $user->id())
+      ->condition('privileges.status', 1);
+
+    $query->addExpression("DATE_FORMAT(field_start_at.field_start_at_value, '%Y-%m-%d')", 'formated_day');
+
+    // Order privilege from low to high permissions.
+    // MySQL only.
+    $query->addExpression("find_in_set(privilege, 'activity_members,activity_maintainers,activity_organizers')", 'order_privileges');
+    $query->orderBy('formated_day', 'ASC');
+    $query->orderBy('order_privileges');
+
+    $rows = $query->execute()->fetchAll();
+
+    $privileges = [];
+    foreach ($rows as $row) {
+      $privileges[$row->formated_day][$row->privilege] = $row->privilege;
     }
 
     return $privileges;